# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/privacy/me": {
      "get": {
        "operationId": "PrivacyController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Privacidade"
        ]
      }
    },
    "/api/v1/privacy/me/requests": {
      "get": {
        "operationId": "PrivacyController_mine",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Privacidade"
        ]
      },
      "post": {
        "operationId": "PrivacyController_request",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrivacyRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Privacidade"
        ]
      }
    },
    "/api/v1/privacy/me/consents": {
      "post": {
        "operationId": "PrivacyController_consent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Privacidade"
        ]
      }
    },
    "/api/v1/privacy/me/export": {
      "get": {
        "operationId": "PrivacyController_ownExport",
        "parameters": [
          {
            "name": "includeFinancial",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Privacidade"
        ]
      }
    },
    "/api/v1/privacy/members/{id}/export": {
      "get": {
        "operationId": "PrivacyController_export",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeFinancial",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Privacidade"
        ]
      }
    },
    "/api/v1/privacy/requests": {
      "get": {
        "operationId": "PrivacyController_requests",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100000,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 120,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Privacidade"
        ]
      }
    },
    "/api/v1/privacy/requests/{id}": {
      "patch": {
        "operationId": "PrivacyController_resolve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolvePrivacyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Privacidade"
        ]
      }
    },
    "/api/v1/privacy/retention": {
      "get": {
        "operationId": "PrivacyController_policies",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Privacidade"
        ]
      },
      "post": {
        "operationId": "PrivacyController_policy",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetentionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Privacidade"
        ]
      }
    },
    "/api/v1/privacy/members/{id}/legal-hold": {
      "patch": {
        "operationId": "PrivacyController_hold",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LegalHoldDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Privacidade"
        ]
      }
    },
    "/api/v1/privacy/members/{id}/anonymize": {
      "post": {
        "operationId": "PrivacyController_anonymize",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnonymizeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Privacidade"
        ]
      }
    },
...
  },
  "info": {
    "title": "Ponte API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "Object": {
        "type": "object",
        "properties": {}
      },
      "PrivacyRequestDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "EXPORT",
              "RECTIFICATION",
              "ANONYMIZATION"
            ]
          },
          "details": {
            "type": "string",
            "minLength": 12,
            "maxLength": 2000
          }
        },
        "required": [
          "kind",
          "details"
        ]
      },
      "ConsentDto": {
        "type": "object",
        "properties": {
          "purpose": {
            "type": "string",
            "minLength": 3,
            "maxLength": 120
          },
          "policyVersion": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "granted": {
            "type": "boolean"
          }
        },
        "required": [
          "purpose",
          "policyVersion",
          "granted"
        ]
      },
      "ResolvePrivacyDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "COMPLETED",
              "REJECTED"
            ]
          },
          "resolution": {
            "type": "string",
            "minLength": 12,
            "maxLength": 2000
          }
        },
        "required": [
          "status",
          "resolution"
        ]
      },
      "RetentionDto": {
        "type": "object",
        "properties": {
          "dataClass": {
            "type": "string",
            "enum": [
              "MEMBER_PROFILE",
              "FINANCIAL_RECORD",
              "AUDIT_LOG",
              "BANK_FILE",
              "CONSENT"
            ]
          },
          "minimumMonths": {
            "type": "number",
            "minimum": 0,
            "maximum": 1200
          },
          "legalBasis": {
            "type": "string",
            "minLength": 12,
            "maxLength": 2000
          }
        },
        "required": [
          "dataClass",
          "minimumMonths",
          "legalBasis"
        ]
      },
      "LegalHoldDto": {
        "type": "object",
        "properties": {
          "legalHold": {
            "type": "boolean"
          }
        },
        "required": [
          "legalHold"
        ]
      },
      "AnonymizeDto": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "requestId"
        ]
      },
...
    }
  }
}